+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+## 1.1.8 (2019-09-08)
+ * Styling sidebar submenu
+
 ## 1.1.7 (2019-09-07)
  * Added bug report functionality
  * Fix check updates for modules. Add method getAppVersion() for layout. Update dependencies.